refactor: 사용자의 크루, 게스트 관련 조회 로직에 대해 구현계층 적용 및 각 도메인으로 로직 이관
