+++ src/main/java/kr/pickple/back/crew/dto/response/CrewMemberRegistrationStatusResponse.java
@@ -1,13 +1,12 @@
-package kr.pickple.back.member.dto.response;
+package kr.pickple.back.crew.dto.response;
 
 import kr.pickple.back.common.domain.RegistrationStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
-@AllArgsConstructor(staticName = "of")
-public class GameMemberRegistrationStatusResponse {
+@AllArgsConstructor(staticName = "from")
+public class CrewMemberRegistrationStatusResponse {
 
     private final RegistrationStatus memberRegistrationStatus;
-    private final Boolean isReviewDone;
 }
